+++ src/database/queries/userExtension.js
@@ -263,10 +263,19 @@ module.exports = class MenteeExtensionQueries {
 				raw: true,
 			}
 
-			const result = unscoped
+			let result = unscoped
 				? await MenteeExtension.unscoped().findAll(query)
 				: await MenteeExtension.findAll(query)
 
+			if (options & options.attributes && options.attributes.includes('email')) {
+				await Promise.all(
+					result.map(async (userInfo) => {
+						if (userInfo && userInfo.email) {
+							userInfo.email = await emailEncryption.decrypt(userInfo.email.toLowerCase())
+						}
+					})
+				)
+			}
 			return result
 		} catch (error) {
 			console.log(error)
